backend/requirements.txt
flask
flask-socketio
torch
transformers
requests
gitpython
sqlite3
